NeighborlyAPI/requirements.txt
azure-functions
pymongo==3.10.1
